Add seed file for DB population and POST request for new loans  (#14)
This PR introduces the following key changes:

1. Seed File for Database Population
Added a seed file to document and automate the initial database setup.
This file provides details on how the database was populated, making it easier for future modifications and testing.

2. Allow Loan Addition Without Risk Score
Banks will continuously add new loans, some of which may not have an assigned risk score.
The system now allows loans to be added without a risk score.
The ML model will automatically generate a risk score for these loans before they proceed to tranching.